Move admin elements into one folder.
