+++ src/ims/element/admin/root/__init__.py
@@ -15,22 +15,12 @@
 ##
 
 """
-Admin access control page.
+Admin page element.
 """
 
-from ims.config import Configuration
-
-from .._page import Page
+from ._admin import AdminPage
 
 
-__all__ = ()
-
-
-
-class AdminAccessControlPage(Page):
-    """
-    Admin access control page.
-    """
-
-    def __init__(self, config: Configuration) -> None:
-        super().__init__(config=config, title="Edit Event Access Control")
+__all__ = (
+    "AdminPage",
+)
